requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
cffi==1.15.0
cryptography==36.0.2
dj-database-url==0.5.0
Django==4.0.3
django-on-heroku==1.1.2
django-rest-knox==4.2.0
djangorestframework==3.13.1
gunicorn==20.1.0
psycopg2-binary==2.9.3
pycparser==2.21
pytz==2022.1
sqlparse==0.4.2
whitenoise==6.0.0
